I04_1-156: moved some overlays to utils
