+++ dls_barcode/camera/capture_worker.py
@@ -1,7 +1,7 @@
 import time
 import queue
 
-from .overlay import Overlay
+from dls_util.image import Overlay
 from .stream_action import StreamAction
 from dls_util.image import Image
 from dls_util.cv import CameraStream
