+++ dls_util/image/__init__.py
@@ -1,2 +1,3 @@
 from .color import Color
 from .image import Image
+from .overlay import Overlay
